added linear scaling for stars instead of relative buckets

+++ MainScreen/Assets/Scripts/Folder.cs
@@ -9,7 +9,7 @@ public class Folder : MonoBehaviour {
     public Dictionary<string, GameObject> children = new Dictionary<string, GameObject>();
     public SpringJoint spring;
     public Rigidbody rb;
-    public int size;
+    public float size;
     public int lastModifiedDate;
     public Text text;
     public float extraglow = 0;
